+++ markwon-core/src/main/java/io/noties/markwon/utils/LeadingMarginUtils.java
@@ -4,7 +4,6 @@ import android.text.Spanned;
 
 public abstract class LeadingMarginUtils {
 
-    @SuppressWarnings("BooleanMethodIsAlwaysInverted")
     public static boolean selfStart(int start, CharSequence text, Object span) {
         return text instanceof Spanned && ((Spanned) text).getSpanStart(span) == start;
     }